requirements.txt
opencv-python
SciPy
NumPy == 1.20.0
tensorflow
streamlit==1.22.0
streamlit-extras
pillow == 10.4.0
datetime